+++ resource/opentelemetry-resource-detector-azure/src/opentelemetry/resource/detector/azure/vm.py
@@ -19,11 +19,14 @@ from urllib.request import Request, urlopen
 from urllib.error import URLError
 
 from opentelemetry.sdk.resources import ResourceDetector, Resource
-from opentelemetry.semconv.resource import ResourceAttributes, CloudPlatformValues, CloudProviderValues
+from opentelemetry.semconv.resource import (
+    ResourceAttributes,
+    CloudPlatformValues,
+    CloudProviderValues,
+)
 
 
 # TODO: Remove when cloud resource id is no longer missing in Resource Attributes
-_CLOUD_RESOURCE_ID_RESOURCE_ATTRIBUTE = "cloud.resource_id"
 _AZURE_VM_METADATA_ENDPOINT = "http://169.254.169.254/metadata/instance/compute?api-version=2021-12-13&format=json"
 _AZURE_VM_SCALE_SET_NAME_ATTRIBUTE = "azure.vm.scaleset.name"
 _AZURE_VM_SKU_ATTRIBUTE = "azure.vm.sku"
@@ -35,7 +38,7 @@ EXPECTED_AZURE_AMS_ATTRIBUTES = [
     ResourceAttributes.CLOUD_PLATFORM,
     ResourceAttributes.CLOUD_PROVIDER,
     ResourceAttributes.CLOUD_REGION,
-    _CLOUD_RESOURCE_ID_RESOURCE_ATTRIBUTE,
+    ResourceAttributes.CLOUD_RESOURCE_ID,
     ResourceAttributes.HOST_ID,
     ResourceAttributes.HOST_NAME,
     ResourceAttributes.HOST_TYPE,
@@ -81,7 +84,7 @@ class _AzureVMMetadataServiceRequestor:
             ams_value = CloudProviderValues.AZURE.value
         elif attribute_key == ResourceAttributes.CLOUD_REGION:
             ams_value = metadata_json["location"]
-        elif attribute_key == _CLOUD_RESOURCE_ID_RESOURCE_ATTRIBUTE:
+        elif attribute_key == ResourceAttributes.CLOUD_RESOURCE_ID:
             ams_value = metadata_json["resourceId"]
         elif attribute_key == ResourceAttributes.HOST_ID or \
             attribute_key == ResourceAttributes.SERVICE_INSTANCE_ID: